chore(channel): merge all channel stuffs
