--- channel/router/README.md
@@ -1,10 +1,0 @@
-# walrus-channel-router
-
-Channel configuration and routing for Walrus.
-
-Provides `ChannelConfig`, `build_router`, and channel spawn helpers for
-wiring multiple platform channels into the Walrus daemon event loop.
-
-## License
-
-GPL-3.0
