--- channel/telegram/README.md
@@ -1,10 +1,0 @@
-# walrus-telegram
-
-Telegram Bot API channel adapter for Walrus.
-
-Implements the `Channel` trait via `TelegramChannel` using direct HTTP
-requests to the Telegram Bot API (no teloxide dependency).
-
-## License
-
-GPL-3.0
