+++ crates/channel/src/channel.rs
@@ -1,78 +1,13 @@
-//! Channel abstraction for the walrus protocol.
-//!
-//! Defines the [`Channel`] trait, platform types, message types, routing,
-//! and menu abstractions. Channel implementations (e.g. Telegram) live in
-//! separate crates and implement [`Channel`] to produce a [`ChannelHandle`].
+//! Channel trait and live connection handle types.
 
+use crate::message::{ChannelMessage, Platform};
 use anyhow::Result;
-use compact_str::CompactString;
 use std::{future::Future, pin::Pin, sync::Arc};
 use tokio::sync::mpsc;
 
 /// Type-erased async sender function.
 type SenderFn =
     Arc<dyn Fn(ChannelMessage) -> Pin<Box<dyn Future<Output = Result<()>> + Send>> + Send + Sync>;
-
-pub mod router;
-
-pub use router::{ChannelRouter, RoutingRule, parse_platform};
-
-/// Messaging platform identifier.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum Platform {
-    /// Telegram messaging platform.
-    Telegram,
-}
-
-/// A message received from or sent to a channel.
-#[derive(Debug, Clone)]
-pub struct ChannelMessage {
-    /// Platform this message belongs to.
-    pub platform: Platform,
-    /// Channel/chat identifier on the platform.
-    pub channel_id: CompactString,
-    /// Sender identifier on the platform.
-    pub sender_id: CompactString,
-    /// Message text content.
-    pub content: String,
-    /// Attached files or media.
-    pub attachments: Vec<Attachment>,
-    /// ID of the message being replied to, if any.
-    pub reply_to: Option<CompactString>,
-    /// Unix timestamp when the message was created.
-    pub timestamp: u64,
-}
-
-/// A file or media attachment.
-#[derive(Debug, Clone)]
-pub struct Attachment {
-    /// Type of attachment.
-    pub kind: AttachmentKind,
-    /// URL or path to the attachment.
-    pub url: String,
-    /// Optional human-readable name.
-    pub name: Option<String>,
-}
-
-/// Type of attachment content.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum AttachmentKind {
-    /// Image file (PNG, JPG, etc.).
-    Image,
-    /// Generic file.
-    File,
-    /// Audio file.
-    Audio,
-    /// Video file.
-    Video,
-}
-
-impl From<ChannelMessage> for wcore::model::Message {
-    fn from(msg: ChannelMessage) -> Self {
-        wcore::model::Message::user(msg.content)
-    }
-}
 
 /// A live channel connection that can receive and send messages.
 ///
