+++ crates/daemon/src/config/mod.rs
@@ -4,11 +4,11 @@ use std::collections::BTreeMap;
 
 pub use ::model::{ProviderConfig, ProviderManager};
 use anyhow::Result;
+pub use channel::ChannelConfig;
 use compact_str::CompactString;
 pub use default::{
     AGENTS_DIR, CRON_DIR, DATA_DIR, SKILLS_DIR, global_config_dir, scaffold_config_dir, socket_path,
 };
-pub use router::ChannelConfig;
 use serde::{Deserialize, Serialize};
 pub use {
     agent::AgentConfig,
