--- protocol/channel/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "walrus-channel"
-description = "Channel abstraction for Walrus protocol"
-documentation = "https://docs.rs/walrus-channel"
-readme = "README.md"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-license.workspace = true
-repository.workspace = true
-keywords.workspace = true
-
-[dependencies]
-wcore.workspace = true
-
-# crates.io
-anyhow.workspace = true
-compact_str.workspace = true
-futures-core.workspace = true
-serde.workspace = true
-tokio.workspace = true
