--- protocol/channel/README.md
@@ -1,10 +1,0 @@
-# walrus-pchannel
-
-Channel abstraction for the Walrus platform.
-
-Provides the `Channel` trait, `Platform` enum, `ChannelMessage`, and
-`ChannelRouter` for routing messages across multiple communication platforms.
-
-## License
-
-GPL-3.0
